Split app.py into app.py, controller.py and utils.py
